safari-534.53-branch/LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.layoutTestController) {
  layoutTestController.dumpAsText();
  layoutTestController.dumpChildFramesAsText();
}
</script>
</head>
<body>
  <p>
    None of these scripts should execute even though there are parse errors in the policy.
  </p>
<script>
function test(shouldRun, policy) {
    document.write('<iframe src="http://127.0.0.1:8000/security/contentSecurityPolicy/resources/echo-script-src.pl?should_run=');
    document.write(shouldRun);
    document.write('&q=http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script.js&csp=');
    document.write(escape(policy));
    document.write('"></iframe>\n');
}


test('yes', 'script-src 127.0.0.1:8000');
test('yes', 'script-src http://127.0.0.1:8000');
test('no',  'script-src https://127.0.0.1:8000');
test('no',  'script-src https://127.0.0.1:*');
test('no',  'script-src https://127.0.0.1:*  ');
test('no',  'script-src https://127.?.0.1:*');
test('yes', 'script-src https://127.0.0.1:\t*  ');
test('yes', 'script-src\thttp://127.0.0.1:8000');
test('yes', 'script-src http://127.0.0.1:8000   \t ');
test('yes', 'script-src http://127.0.0.1:* ');
test('yes', 'script-src http:');
test('no',  'script-src https:');
test('no',  'script-src    https:  ');
test('no',  'script-src    https://');
test('no',  'script-src https:  taco');
test('yes', 'script-src https:  \x08');
test('no',  'script-src \'none\'');
test('yes', '\t\t\tscript-src    http://127.0.0.1:8000  \t\t  https:');

</script>
</body>
</html>
